requirements.txt
python-telegram-bot[job-queue]>=21.0
httpx>=0.27.0
matplotlib>=3.10.0
pandas>=2.2.0
yfinance>=0.2.50
python-dotenv>=1.0.0
aiosqlite>=0.20.0
